package-refactor: Fix deprecations move page specs into package
