+++ app/controllers/RequestDataController.scala
@@ -49,7 +49,7 @@ class RequestDataController @Inject() (
         case _           =>
           logErrorsAndContinue(
             "Email was not found",
-            routes.IndexController.onPageLoad
+            routes.IndexController.onPageLoad()
           )
       }
   }
