+++ build.sbt
@@ -11,7 +11,7 @@ ThisBuild / scalaVersion := "2.13.12"
 lazy val microservice = (project in file("."))
   .enablePlugins(PlayScala, SbtDistributablesPlugin)
   .disablePlugins(JUnitXmlReportPlugin) //Required to prevent https://github.com/scalatest/scalatest/issues/1427
-  .settings(inConfig(Test)(testSettings): _*)
+  .settings(inConfig(Test)(testSettings) *)
   .settings(ThisBuild / useSuperShell := false)
   .settings(
     name := appName,
@@ -59,7 +59,7 @@ lazy val microservice = (project in file("."))
     Assets / pipelineStages := Seq(concat)
   )
 
-lazy val testSettings: Seq[Def.Setting[_]] = Seq(
+lazy val testSettings: Seq[Def.Setting[?]] = Seq(
   fork := true,
   unmanagedSourceDirectories += baseDirectory.value / "test-utils",
   scalafmtOnCompile := true
@@ -74,3 +74,5 @@ lazy val it =
 addCommandAlias("testAndCoverage", ";clean;coverage;test;it/test;coverageReport")
 addCommandAlias("prePR", ";scalafmt;test:scalafmt;testAndCoverage")
 addCommandAlias("preMerge", ";scalafmtCheckAll;testAndCoverage")
+
+scalacOptions += "-deprecation"
